refactor(examples): restructure CrewAI examples to standard crewai create layout
Refactor all 4 CrewAI examples (steering_financial_agent, evaluator_showcase,
secure_research_crew, content_publishing_flow) from single-file monoliths to
the standard `crewai create crew/flow` project structure:

- src/<package>/main.py — entry point, scenarios, verify_setup
- src/<package>/crew.py — @CrewBase class with YAML config (crew examples)
- src/<package>/config/agents.yaml + tasks.yaml — agent/task definitions
- src/<package>/tools/*.py — tool implementations with @control() wrappers
- pyproject.toml — [tool.crewai] type, [project.scripts] entry points

All examples tested and passing with identical behavior.
