+++ examples/crewai/evaluator_showcase/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "agent-control-crewai-evaluator-showcase"
+name = "evaluator-showcase"
 version = "0.1.0"
 description = "CrewAI agent demonstrating all built-in evaluators: regex, list, json, and sql"
 requires-python = ">=3.12"
@@ -17,9 +17,16 @@ dev = [
     "pytest-asyncio>=0.23.0",
 ]
 
+[project.scripts]
+evaluator_showcase = "evaluator_showcase.main:main"
+run_crew = "evaluator_showcase.main:main"
+
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-include = ["*.py"]
+packages = ["src/evaluator_showcase"]
+
+[tool.crewai]
+type = "crew"
