+++ examples/crewai/evaluator_showcase/src/evaluator_showcase/main.py
@@ -17,7 +17,7 @@ PREREQUISITE:
 
     Then run this example:
 
-        $ uv run python data_analyst.py
+        $ uv run evaluator_showcase
 
 Scenarios:
     1. Safe SELECT query            -> SQL evaluator ALLOWS
@@ -30,14 +30,13 @@ Scenarios:
     8. Missing purpose field        -> JSON evaluator STEERS (then allowed)
 """
 
-import asyncio
 import json
 import os
 
 import agent_control
-from agent_control import ControlSteerError, ControlViolationError, control
-from crewai import Agent, Crew, LLM, Task
-from crewai.tools import tool
+
+from evaluator_showcase.tools import create_sql_tool, create_analysis_tool
+from evaluator_showcase.crew import EvaluatorShowcaseCrew
 
 # ── Configuration ───────────────────────────────────────────────────────
 AGENT_NAME = "crewai-data-analyst"
@@ -48,203 +47,6 @@ agent_control.init(
     agent_description="CrewAI data analyst with all evaluator types",
     server_url=SERVER_URL,
 )
-
-
-# ── Simulated Database ──────────────────────────────────────────────────
-# In a real app these would hit an actual database. The simulated responses
-# let us demonstrate how Agent Control inspects both input AND output.
-
-SIMULATED_RESULTS = {
-    "safe": (
-        "| order_id | product   | total  |\n"
-        "|----------|-----------|--------|\n"
-        "| 1001     | Widget A  | 29.99  |\n"
-        "| 1002     | Widget B  | 49.99  |\n"
-        "| 1003     | Gadget X  | 149.99 |\n"
-        "\n3 rows returned."
-    ),
-    "pii": (
-        "| customer_id | name       | ssn         | email              |\n"
-        "|-------------|------------|-------------|--------------------|  \n"
-        "| C-101       | John Smith | 123-45-6789 | john@example.com   |\n"
-        "| C-102       | Jane Doe   | 987-65-4321 | jane@example.com   |\n"
-        "\n2 rows returned."
-    ),
-}
-
-
-# ── Tool 1: SQL Query Runner ────────────────────────────────────────────
-
-def create_sql_tool():
-    """Build the SQL query tool with @control protection."""
-
-    async def _run_sql_query(query: str) -> str:
-        """Execute a SQL query against the database (protected)."""
-        # Simulate: if query touches customers_full, return PII results
-        if "customers_full" in query.lower():
-            return SIMULATED_RESULTS["pii"]
-        return SIMULATED_RESULTS["safe"]
-
-    _run_sql_query.name = "run_sql_query"  # type: ignore[attr-defined]
-    _run_sql_query.tool_name = "run_sql_query"  # type: ignore[attr-defined]
-    controlled_fn = control()(_run_sql_query)
-
-    @tool("run_sql_query")
-    def run_sql_query_tool(query: str) -> str:
-        """Run a SQL query against the company database.
-
-        Args:
-            query: The SQL query to execute
-        """
-        if isinstance(query, dict):
-            query = query.get("query", str(query))
-
-        print(f"\n  [SQL TOOL] Query: {query[:80]}...")
-
-        try:
-            result = asyncio.run(controlled_fn(query=query))
-            print(f"  [SQL TOOL] Query executed successfully")
-            return result
-
-        except ControlViolationError as e:
-            print(f"  [SQL TOOL] BLOCKED by {e.control_name}: {e.message[:100]}")
-            return f"QUERY BLOCKED: {e.message}"
-
-        except Exception as e:
-            print(f"  [SQL TOOL] Error: {e}")
-            return f"Query error: {e}"
-
-    return run_sql_query_tool
-
-
-# ── Tool 2: Data Analyzer ───────────────────────────────────────────────
-
-def create_analysis_tool():
-    """Build the analysis tool with JSON validation and steering."""
-
-    # Defer LLM creation -- only needed if OPENAI_API_KEY is set
-    llm = None
-    if os.getenv("OPENAI_API_KEY"):
-        llm = LLM(model="gpt-4o-mini", temperature=0.3)
-
-    async def _analyze_data(request: dict) -> str:
-        """Run data analysis (protected by JSON validation controls).
-
-        Takes a single dict param so the @control() decorator sends it
-        as input.request — and the JSON evaluator can check which fields
-        are present or absent.
-        """
-        dataset = request.get("dataset", "")
-        date_range = request.get("date_range", "")
-        max_rows = request.get("max_rows", 1000)
-        purpose = request.get("purpose", "")
-
-        prompt = f"""Summarize this data analysis in 2-3 sentences:
-- Dataset: {dataset}
-- Date range: {date_range}
-- Max rows: {max_rows}
-- Purpose: {purpose}
-
-Provide a brief, professional analysis summary."""
-
-        return llm.call([{"role": "user", "content": prompt}])
-
-    _analyze_data.name = "analyze_data"  # type: ignore[attr-defined]
-    _analyze_data.tool_name = "analyze_data"  # type: ignore[attr-defined]
-    controlled_fn = control()(_analyze_data)
-
-    @tool("analyze_data")
-    def analyze_data_tool(request: str) -> str:
-        """Analyze a dataset with validation controls.
-
-        Args:
-            request: JSON string with fields: dataset (required), date_range (required),
-                max_rows (optional, 1-10000), purpose (recommended for audit compliance)
-        """
-        if isinstance(request, dict):
-            params = request
-        else:
-            try:
-                params = json.loads(request)
-            except (json.JSONDecodeError, TypeError):
-                return f"Invalid request format. Expected JSON, got: {request!r}"
-
-        # Build the request dict — only include fields that have values.
-        # The JSON evaluator checks which fields are PRESENT in this dict,
-        # so omitting a field triggers the "required_fields" check.
-        request_dict: dict = {}
-        if params.get("dataset"):
-            request_dict["dataset"] = params["dataset"]
-        if params.get("date_range"):
-            request_dict["date_range"] = params["date_range"]
-        if params.get("max_rows") is not None:
-            request_dict["max_rows"] = int(params["max_rows"])
-        if params.get("purpose"):
-            request_dict["purpose"] = params["purpose"]
-
-        print(f"\n  [ANALYSIS TOOL] Request: {request_dict}")
-
-        # Steering retry loop
-        max_attempts = 3
-        for attempt in range(1, max_attempts + 1):
-            try:
-                result = asyncio.run(controlled_fn(request=request_dict))
-                print(f"  [ANALYSIS TOOL] Analysis complete")
-                return result
-
-            except ControlViolationError as e:
-                print(f"  [ANALYSIS TOOL] BLOCKED by {e.control_name}: {e.message[:100]}")
-                return f"ANALYSIS BLOCKED: {e.message}"
-
-            except ControlSteerError as e:
-                print(f"  [ANALYSIS TOOL] STEERED by {e.control_name}")
-                try:
-                    guidance = json.loads(e.steering_context)
-                except (json.JSONDecodeError, TypeError):
-                    guidance = {}
-
-                reason = guidance.get("reason", "Correction needed")
-                actions = guidance.get("required_actions", [])
-                print(f"    Reason: {reason}")
-                print(f"    Actions: {actions}")
-
-                if "collect_purpose" in actions:
-                    auto_purpose = f"Quarterly {request_dict.get('dataset', 'data')} analysis for business reporting"
-                    request_dict["purpose"] = auto_purpose
-                    print(f"    Auto-filled purpose: {auto_purpose}")
-
-                continue
-
-        return "ANALYSIS FAILED: Could not satisfy all controls."
-
-    return analyze_data_tool
-
-
-# ── CrewAI Crew ─────────────────────────────────────────────────────────
-
-def create_analyst_crew(sql_tool, analysis_tool):
-    analyst = Agent(
-        role="Data Analyst",
-        goal="Execute data queries and analysis while respecting all data governance controls",
-        backstory=(
-            "You are a data analyst at a company with strict data governance policies. "
-            "You use run_sql_query to query the database and analyze_data for analysis. "
-            "You always comply with security controls and never attempt to bypass them."
-        ),
-        tools=[sql_tool, analysis_tool],
-        verbose=True,
-    )
-
-    task = Task(
-        description=(
-            "Execute this data request: {request}\n\n"
-            "Use the appropriate tool and report the outcome."
-        ),
-        expected_output="Query results or analysis, or an explanation if blocked by controls",
-        agent=analyst,
-    )
-
-    return Crew(agents=[analyst], tasks=[task], verbose=True)
 
 
 # ── Scenario Runner ─────────────────────────────────────────────────────
@@ -498,7 +300,7 @@ def main():
         print("  Agent autonomously handles a multi-step data request")
         print("#" * 60)
 
-        crew = create_analyst_crew(sql_tool, analysis_tool)
+        crew = EvaluatorShowcaseCrew().crew()
 
         print("\n  Running crew with a safe data request...")
         result = crew.kickoff(
