+++ examples/crewai/secure_research_crew/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "agent-control-crewai-research-crew"
+name = "secure-research-crew"
 version = "0.1.0"
 description = "CrewAI multi-agent research crew with per-agent policies via Agent Control"
 requires-python = ">=3.12"
@@ -11,9 +11,16 @@ dependencies = [
     "python-dotenv>=1.0.0",
 ]
 
+[project.scripts]
+secure_research_crew = "secure_research_crew.main:main"
+run_crew = "secure_research_crew.main:main"
+
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-include = ["*.py"]
+packages = ["src/secure_research_crew"]
+
+[tool.crewai]
+type = "crew"
