+++ examples/crewai/steering_financial_agent/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "agent-control-crewai-steering-example"
+name = "steering-financial-agent"
 version = "0.1.0"
 description = "CrewAI financial agent with steering, deny, and warn actions via Agent Control"
 requires-python = ">=3.12"
@@ -11,15 +11,16 @@ dependencies = [
     "python-dotenv>=1.0.0",
 ]
 
-[project.optional-dependencies]
-dev = [
-    "pytest>=8.0.0",
-    "pytest-asyncio>=0.23.0",
-]
+[project.scripts]
+steering_financial_agent = "steering_financial_agent.main:main"
+run_crew = "steering_financial_agent.main:main"
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-include = ["*.py"]
+packages = ["src/steering_financial_agent"]
+
+[tool.crewai]
+type = "crew"
